+++ src/AzureAdB2cMsalConsoleApp/AzureAdB2cMsalConsoleApp.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\SampleLib\SampleLib.csproj" />
+  </ItemGroup>
+
 </Project>
